Fix CUDA Architecture for CI with old CMake
The configuration for CI changed to using the newer CMake option
`CMAKE_CUDA_ARCHITECTURES` for turing builds instead of the now
depreciated `VTKm_CUDA_Architecture` setting. However, this new
setting only works for newer versions of CMake, and some of the
CI docker images still have older versions of CMake.

This happened to work on the CI because the CI was run on
machines that had the right CUDA device installed (I guess), but
did not work with the `reproduce_ci_env.py` utility. Fix this
by checking the CMake version first.