Fix Fields and CoordinateSystems of Lagrangian filter outputs
The original versions of Lagrangian and LagrangianStructures mapped the
global fields from inputs to output. Restore that capability.

Also changed the output CoordinateSystem of Lagrangian to be uniform
points instead of rectilinear points. The points were uniformly spaced
anyway, so it is more clear to have uniform points in the coordinate
system.