Enable floating point exceptions on apple
Clang for apple does not support `fpenableexcept`, so make a special way
to turn them on to help developers debug code. This code is based on the
suggestions at:

https://stackoverflow.com/questions/37819235/how-do-you- enable-floating-point-exceptions-for-clang-in-os-x

https://stackoverflow.com/questions/71821666/trapping-floating-point-exceptions-and-signal-handling-on-apple-silicon

This code is actually disabled because it does not work on all known
platforms (most notably one of the dashboard machines). However, I
would like to keep the code there to assist debugging.