Update docs CI to match job naming convention
The naming convention for CI jobs is generally `stage:name` where the
`stage` matches the GitLab stage and the `name` is a description.
However, the documentation jobs broke this convention by being called
`docs:type` where `type` is `continuous`, `latest`, or `master`. This
breaks the convention because `docs` is not a stage and `type` does not
really describe the build. This fouls up things like
`reproduce_ci_env.py` that expect the first part of the name to be the
same as the stage.

This changes the names to `build:docs-continuous`, `build:docs-latest`,
and `build:docs-master`. This follows the convention of matching the
names with the stage.