Add guide information for accessing/allocating data in arrays
In the previous guide, this was (mostly) in its own chapter. This may
have been an excessive amount of splitting things up.

There is still a missing section on the interface to the execution
environment (i.e. PrepareFor*). I plan to put that in the execution
objects chapter as this is the most likely place for a user to use them.