Remove unused method from MIR tables
The implementation of this method was incorrect as it referenced a class
member that does not exist. Many compilers allowed it in a templated
class when the method was never used, but other compilers attempt to
compile the inlined method regardless.

Since the method clearly is not needed, the easy solution is to remove
it.