Change the README to reference an example
Previously, the README.md file listed a simple example to demonstrate
what Viskores code looks like. However, it was pointed out that the
code was out of date (using deprecated features).

To prevent the example getting out of date, change the listing to a
reference to the code in the examples directory. When viewing the
README on GitHub, users will be able to simply click the link and view
this file directly. Because the code is compiled as part of the CI,
we don't have to worry about it being out of date.