Update use of Actor rendering class
When the rendering classes were originally created, `Actor` accepted
the cell set, coordinate system, and field independently. More
recently (but not that recently), new constructors were made that
accept a `DataSet` object and pulls these out for you. This is generally
a better way to construct actors, so update the documentation and
testing code to follow this convention.

Also added new `Actor` constructors that do not require passing
the name of the coordinate system. This is for the common use case
where each dataset has exactly one coordinate system.